requirements.txt
pandas==2.3.3
numpy==2.4.10
matplotlib==3.10.8
seaborn==0.13.2
opencv-python==4.12.0.88
tqdm==4.67.1
pillow==11.3.0
scikit-image==0.25.2
scikit-learn==1.8.0
tensorflow==2.20.0
tensorflowjs==4.10.0
importlib_metadata==8.7.0
Flask==3.1.2